requirements.txt
fastapi==0.75.1
uvicorn==0.17.6
sqlalchemy==1.4.35
python-dotenv==0.20.0
python-multipart==0.0.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4